pyproject.toml
[tool.poetry]
name = "workflow-iimi"
version = "0.1.0"
description = ""
authors = ["Ian Boyes"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
virtool-workflow = {version="^6.0.0a7", allow-prereleases=true}
biopython = "^1.81"
pydantic-factories = "^1.17.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.23.2"
pytest-aiohttp = "^1.0.5"
syrupy = "^4.6.0"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = ["asyncio_mode: auto"]
